fix(pilote-ppg-auth): Cache-Control no-store + casing x-forwarded-host + entrypoint direct
- Extrait createApp dans app.ts, server.ts devient un entrypoint qui lance direct (plus de guard import.meta.url peu fiable).
- Ajoute Cache-Control: no-store sur la réponse du challenge ACME (aligné avec l'endpoint Next.js).
- Override x-forwarded-host en lowercase pour éviter la duplication avec la clé venant du spread de c.req.header() qui sort les clés en lowercase.
